fixed a semi-deep bug in the global namestore
the "global" name store was not global to the elaboration, leading to
problems with reabstraction. elaboration would instantiate at a
generated var, then reabstract over that, and this caused problems when
that same var was used later b/c the abstractions were global only to a
type checker, not global to the whole elaborator, leading to unintended
reuse and thus unintended abstractions. this wasn't an issue in the
non-dependent-implicit versions, but they've been changed as well to
keep things consistent.