replace occurs-check undefined-based instantiation w/ name-based
using `Var undefined` during occurs check was causing mysterious errors
in the Module variant (and possibly other variants, w/o being noticed).
the `undefined` was at some point being inspected, leading to errors.
replacing it with the new `descope` ensures this never happens.
`descope` can only be used in contexts where the named-variable
identity won't ever matter, however. if this causes issues in the
future (and it may!), the final option would be to make occurs check
use the name-store and generate new variables. hopefully this wont be
necessary.